test/requirements.txt
numpy~=1.22.0
opencv-python~=4.7.0.72
scikit-image>=0.20.0
Pillow~=9.5.0
pytest~=7.3.0
pytest-cov~=4.0.0
realesrgan-ncnn-py
